+++ src/depth_limit.rs
@@ -12,29 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{Error, Decode, Input};
+use crate::{Decode, Error, Input};
 
 /// The error message returned when depth limit is reached.
 const DECODE_MAX_DEPTH_MSG: &str = "Maximum recursion depth reached when decoding";
 
 /// Extension trait to [`Decode`] for decoding with a maximum recursion depth.
 pub trait DecodeLimit: Sized {
-	/// Decode `Self` with the given maximum recursion depth.
+	/// Decode `Self` with the given maximum recursion depth and advance `input` by the number of
+	/// bytes consumed.
 	///
 	/// If `limit` is hit, an error is returned.
-	fn decode_with_depth_limit(limit: u32, input: &[u8]) -> Result<Self, Error>;
-
-	/// Decode `Self` and advance `input` by the number of bytes consumed.
-	///
-	/// If `limit` is hit, an error is returned.
-	fn decode_and_advance_with_depth_limit<I: Input>(limit: u32, input: &mut I) -> Result<Self, Error>;
+	fn decode_with_depth_limit<I: Input>(limit: u32, input: &mut I) -> Result<Self, Error>;
 
 	/// Decode `Self` and consume all of the given input data.
 	///
 	/// If not all data is consumed or `limit` is hit, an error is returned.
-	fn decode_all_with_depth_limit(limit: u32, input: &[u8]) -> Result<Self, Error>;
+	fn decode_all_with_depth_limit(limit: u32, input: &mut &[u8]) -> Result<Self, Error>;
 }
-
 
 struct DepthTrackingInput<'a, I> {
 	input: &'a mut I,
@@ -42,7 +37,7 @@ struct DepthTrackingInput<'a, I> {
 	max_depth: u32,
 }
 
-impl<'a, I:Input> Input for DepthTrackingInput<'a, I> {
+impl<'a, I: Input> Input for DepthTrackingInput<'a, I> {
 	fn remaining_len(&mut self) -> Result<Option<usize>, Error> {
 		self.input.remaining_len()
 	}
@@ -72,36 +67,18 @@ impl<'a, I:Input> Input for DepthTrackingInput<'a, I> {
 }
 
 impl<T: Decode> DecodeLimit for T {
-	fn decode_all_with_depth_limit(limit: u32, input: &[u8]) -> Result<Self, Error> {
-		let mut input = DepthTrackingInput {
-			input: &mut &input[..],
-			depth: 0,
-			max_depth: limit,
-		};
-		let res = T::decode(&mut input)?;
+	fn decode_all_with_depth_limit(limit: u32, input: &mut &[u8]) -> Result<Self, Error> {
+		let t = <Self as DecodeLimit>::decode_with_depth_limit(limit, input)?;
 
-		if input.input.is_empty() {
-			Ok(res)
+		if input.is_empty() {
+			Ok(t)
 		} else {
 			Err(crate::decode_all::DECODE_ALL_ERR_MSG.into())
 		}
 	}
 
-	fn decode_and_advance_with_depth_limit<I: Input>(limit: u32, input: &mut I) -> Result<Self, Error> {
-		let mut input = DepthTrackingInput {
-			input,
-			depth: 0,
-			max_depth: limit,
-		};
-		T::decode(&mut input)
-	}
-
-	fn decode_with_depth_limit(limit: u32, input: &[u8]) -> Result<Self, Error> {
-		let mut input = DepthTrackingInput {
-			input: &mut &input[..],
-			depth: 0,
-			max_depth: limit,
-		};
+	fn decode_with_depth_limit<I: Input>(limit: u32, input: &mut I) -> Result<Self, Error> {
+		let mut input = DepthTrackingInput { input, depth: 0, max_depth: limit };
 		T::decode(&mut input)
 	}
 }
@@ -117,19 +94,38 @@ mod tests {
 		let nested: NestedVec = vec![vec![vec![vec![1]]]];
 		let encoded = nested.encode();
 
-		let decoded = NestedVec::decode_with_depth_limit(3, &encoded).unwrap();
+		let decoded = NestedVec::decode_with_depth_limit(3, &mut encoded.as_slice()).unwrap();
 		assert_eq!(decoded, nested);
-		assert!(NestedVec::decode_with_depth_limit(2, &encoded).is_err());
+		assert!(NestedVec::decode_with_depth_limit(2, &mut encoded.as_slice()).is_err());
 	}
 
 	#[test]
-	fn decode_and_advance_works() {
+	fn decode_limit_advances_input() {
 		type NestedVec = Vec<Vec<Vec<Vec<u8>>>>;
 		let nested: NestedVec = vec![vec![vec![vec![1]]]];
-		let encoded = &mut &nested.encode()[..];
+		let encoded = nested.encode();
+		let encoded_slice = &mut encoded.as_slice();
 
-		let decoded = Vec::<u8>::decode_and_advance_with_depth_limit(1, encoded).unwrap();
+		let decoded = Vec::<u8>::decode_with_depth_limit(1, encoded_slice).unwrap();
 		assert_eq!(decoded, vec![4]);
-		assert!(NestedVec::decode_with_depth_limit(3, encoded).is_err());
+		assert!(NestedVec::decode_with_depth_limit(3, encoded_slice).is_err());
+	}
+
+	#[test]
+	fn decode_all_with_limit_advances_input() {
+		type NestedVec = Vec<Vec<Vec<Vec<u8>>>>;
+		let nested: NestedVec = vec![vec![vec![vec![1]]]];
+		let mut encoded = NestedVec::encode(&nested);
+
+		let decoded = NestedVec::decode_all_with_depth_limit(3, &mut encoded.as_slice()).unwrap();
+		assert_eq!(decoded, nested);
+
+		encoded.extend(&[1, 2, 3, 4, 5, 6]);
+		assert_eq!(
+			NestedVec::decode_all_with_depth_limit(3, &mut encoded.as_slice())
+				.unwrap_err()
+				.to_string(),
+			"Input buffer has still data left after decoding!",
+		);
 	}
 }